# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CBL Trace API DOCUMENTATION",
        "version": "1.0.0"
    },
    "paths": {
        "/api/application-create": {
            "post": {
                "tags": [
                    "CBL Trace API"
                ],
                "summary": "Create",
                "description": "Create Applications",
                "operationId": "uploadFile",
                "consumes": "multipart/form-data",
                "parameters": [{
                    "name": "city[]",
                    "in": "path",
                    "description": "Enter Application Type",
                    "type": "number"
                }, {
                    "name": "name",
                    "in": "path",
                    "description": "Enter Applicant Name",
                    "type": "text"
                },{
                        "name": "phone",
                        "in": "path",
                        "description": "Enter Applicant phone number",
                        "type": "number"
                    },{
                    "name": "address",
                    "in": "path",
                    "description": "Enter Applicant address",
                    "type": "text"
                },{
                    "name": "officeName",
                    "in": "path",
                    "description": "Enter Applicant Office Name",
                    "type": "text"
                },{
                    "name": "officeAddress",
                    "in": "path",
                    "description": "Enter Applicant Office Address",
                    "type": "text"
                },{
                    "name": "designation",
                    "in": "path",
                    "description": "Enter Applicant Designation",
                    "type": "text"
                },{
                    "name": "imgae_upload",
                    "in": "formData",
                    "description": "Applicant Image Upload",
                    "required": false,
                    "type": "file"
                },{
                    "name": "co_applicant_name",
                    "in": "path",
                    "description": "Enter Co-Applicant Name",
                    "type": "text"
                },{
                    "name": "co_applicant_phone",
                    "in": "path",
                    "description": "Enter Co-Applicant Number",
                    "type": "number"
                },{
                    "name": "co_applicant_address",
                    "in": "path",
                    "description": "Enter Co-Applicant Address",
                    "type": "text"
                },{
                    "name": "co_applicant_officeName",
                    "in": "path",
                    "description": "Enter Co-Applicant Office Name",
                    "type": "text"
                },{
                    "name": "co_applicant_designation",
                    "in": "path",
                    "description": "Enter Co-Applicant Designation",
                    "type": "text"
                },{
                    "name": "co_applicant_image",
                    "in": "formData",
                    "description": "Co-Applicant Image Upload",
                    "required": false,
                    "type": "file"
                }
                ],
                "requestBody": {
                    "description": "Application Information provides",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "city[]",
                                    "name",
                                    "phone",
                                    "address",
                                    "officeName",
                                    "officeAddress",
                                    "designation",
                                    "co_applicant_name",
                                    "co_applicant_phone",
                                    "co_applicant_address",
                                    "co_applicant_officeName",
                                    "co_applicant_designation",
                                    "guarantor_name",
                                    "guarantor_phone",
                                    "guarantor_address",
                                    "guarantor_officeName",
                                    "guarantor_officeAddress",
                                    "guarantor_designation",
                                    "second_guarantor_name",
                                    "second_guarantor_phone",
                                    "second_guarantor_address"
                                ],
                                "properties": {
                                    "city[]": {
                                        "type": "string",
                                        "format": "number",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Ratul"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "number",
                                        "example": "01722105963"
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Wasi Tower, House 572/K, Matikata Main Road, Matikata, Dhaka Cantonment, Dhaka"
                                    },
                                    "officeName": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Barikoi Technologies Limited"
                                    },
                                    "officeAddress": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Wasi Tower, House 572/K, Matikata Main Road, Matikata, Dhaka Cantonment, Dhaka"
                                    },
                                    "designation": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Jr. Software Engineer"
                                    },
                                    "co_applicant_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Sadiq"
                                    },
                                    "co_applicant_phone": {
                                        "type": "string",
                                        "format": "number",
                                        "example": "01722105963"
                                    },
                                    "co_applicant_address": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "House 416/2, Iqbal Road, West Shewrapara, Shewrapara, Mirpur, Dhaka"
                                    },
                                    "co_applicant_officeName": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Barikoi Technologies Limited"
                                    },
                                    "co_applicant_officeAddress": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Wasi Tower, House 572/K, Matikata Main Road, Matikata, Dhaka Cantonment, Dhaka"
                                    },
                                    "co_applicant_designation": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Jr. SQA Engineer"
                                    },
                                    "guarantor_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Binoy"
                                    },
                                    "guarantor_phone": {
                                        "type": "string",
                                        "format": "number",
                                        "example": "01681906756"
                                    },
                                    "guarantor_address": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "House 572/K, Matikata Main Road, Matikata, Dhaka Cantonment, Dhaka"
                                    },
                                    "guarantor_officeName": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Barikoi Technologies Limited"
                                    },
                                    "guarantor_officeAddress": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "House 572/K, Matikata Main Road, Matikata, Dhaka Cantonment, Dhaka"
                                    },
                                    "guarantor_designation": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Sales"
                                    },
                                    "second_guarantor_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Sojib"
                                    },
                                    "second_guarantor_phone": {
                                        "type": "string",
                                        "format": "number",
                                        "example": "01971906756"
                                    },
                                    "second_guarantor_address": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "House 416/2, Iqbal Road, West Shewrapara, Shewrapara, Mirpur, Dhaka"
                                    },
                                    "second_guarantor_officeName": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Barikoi Technologies Limited"
                                    },
                                    "second_guarantor_officeAddress": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "House 572/K, Matikata Main Road, Matikata, Dhaka Cantonment, Dhaka"
                                    },
                                    "second_guarantor_designation": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Jr. Software Engineer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "data": {
                                            "application": {
                                                "id": 1533,
                                                "application_id": "CS-4534642"
                                            }
                                        },
                                        "message": "Success",
                                        "status": 200
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "",
                "name": "",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}